funds-manager: custody_client: gas_sponsor: replace allocation w/ fixed dollar value (#366)
