+++ funds-manager/funds-manager-server/src/handlers/mod.rs
@@ -1,0 +1,9 @@
+//! Route handlers for the funds manager
+
+pub mod fee_indexing;
+pub mod gas;
+pub mod hot_wallets;
+pub mod quoters;
+pub mod rpc;
+pub mod swap;
+pub mod vaults;
